requirements.txt
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-heroku==0.3.1
django-oauth-toolkit==1.2.0
django-rest-api==0.1.5
django-rest-auth==0.9.3
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
PyJWT==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
stripe==2.20.3
urllib3==1.24.1
whitenoise==4.1.2